Extract batch to a separated module and make it apply to Facet by default
